Moving methods from delayed js to script js
